refactor(keyframes): split useGraphInteraction into per-surface sub-hooks
Compose useGraphInteraction from four independent sub-hooks — viewport
math and zoom controls, marquee selection with its global pointer
listeners, snap/blocked-range math, and the wheel zoom/pan handler —
sharing the SVG and drag refs through arguments so the orchestrator
hook stays in charge of pointer-move/up dispatch but no longer owns
each surface's state. Drops the orchestrator from 1,194 to 784 LOC
with no behavior change.
